Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,10 +3,10 @@ int main(){
     int a , b , c;
     scanf("%d %d %d",&a,&b,&c);
 
-    int Average;
+    float Average;
     Average = (a + b + c)/3;
 
-    printf("Average: %d\n",Average);
+    printf("Average: %f\n",Average);
     return 0;
 
 }
